+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/anthropic
 
+## 4.0.0-beta.14
+
+### Patch Changes
+
+- Updated dependencies [2e17091]
+  - @ai-sdk/provider-utils@5.0.0-beta.9
+
 ## 4.0.0-beta.13
 
 ### Patch Changes